Fix for ignored shouldCreateOpen options
It seems that at some point the name of newFile and the actual file created during plugin execution got out of sync.  I removed the declaration (const) from the file creation, and changed the name back to newFile, effectively repointing the null reference, which is what the downstream code is looking for in lines 138 through 142.